+++ src/main/java/com/ll/spring_additional/boundedContext/comment/form/CommentForm.java
@@ -16,8 +16,8 @@ public class CommentForm {
 	private Long id;
 	@NotBlank(message = "내용은 필수항목입니다.")
 	private String commentContents;
-	private Integer questionId;
-	private Integer answerId;
+	private Long questionId;
+	private Long answerId;
 	private Boolean secret = false;
 	private Long parentId;
 	private Long commentWriter;
